requirements.txt
attrs==19.3.0             # via pytest
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
idna==2.8                 # via requests
importlib-metadata==1.5.0  # via pluggy, pytest
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyjwt==1.7.1              # via twilio
pyparsing==2.4.6          # via packaging
pytest==5.3.5
python-dotenv==0.12.0
pytz==2019.3              # twilio
requests==2.22.0
six==1.14.0               # via packaging, twilio
twilio==6.35.5
urllib3==1.25.6           # via requests
wcwidth==0.1.8            # via pytest
zipp==2.2.0
